+++ lib/data_layer.ex
@@ -747,6 +747,7 @@ defmodule AshPostgres.DataLayer do
       when type in [:count, :sum, :first, :list, :avg, :max, :min, :exists, :custom],
       do: true
 
+  def can?(_, {:aggregate, :unrelated}), do: true
   def can?(_, :aggregate_filter), do: true
   def can?(_, :aggregate_sort), do: true
   def can?(_, :calculate), do: true
@@ -919,8 +920,12 @@ defmodule AshPostgres.DataLayer do
   end
 
   @impl true
-  def set_tenant(_resource, query, tenant) do
-    {:ok, Map.put(Ecto.Query.put_query_prefix(query, to_string(tenant)), :__tenant__, tenant)}
+  def set_tenant(resource, query, tenant) do
+    if Ash.Resource.Info.multitenancy_strategy(resource) == :context do
+      {:ok, Map.put(Ecto.Query.put_query_prefix(query, to_string(tenant)), :__tenant__, tenant)}
+    else
+      {:ok, query}
+    end
   end
 
   @impl true
@@ -3165,6 +3170,14 @@ defmodule AshPostgres.DataLayer do
         changeset.context
       )
       |> pkey_filter(changeset.data)
+      |> then(fn query ->
+        if changeset.tenant do
+          set_tenant(resource, query, changeset.tenant)
+          |> elem(1)
+        else
+          query
+        end
+      end)
       |> then(fn query ->
         Map.put(
           query,